Create a tile plot of a palette with ggplot

Create a tile plot of a palette with ggplot

Usage

ro_gg_create_palette(df, palette_name, palette)

Arguments

df

dataframe with 3 character columns: number, name and hex

palette_name

Character. Name of the palette to use in plot title

palette

named character vector with hex values to use in scale_fill_manual

Value

ggplot object

See also

Examples

if (FALSE) { # \dontrun{
palette_name <- "categorical"
palette <- ROvis.utils:::get_color_palette(palette_name)
df <- ROvis.utils:::ro_show_colors_get_df(palette)
ROvis.utils:::ro_gg_create_palette(df, palette_name, palette)
} # }